refactor(api): extract data initialization and helpers to utils
